backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data validation
pydantic==2.9.0
pydantic-settings==2.5.0

# OpenAI
openai==1.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Utilities
python-dotenv==1.0.0
